Moved coordinate-related routines into a different file within var module
